# Inline more functions in `vm` crate