--- a/bench/sum/main.sh
+++ b/bench/sum/main.sh
@@ -4,4 +4,4 @@ set -e
 
 cd $(dirname $0)
 
-hyperfine '../../target/release/arachne < main.arc' 'python3 main.py'
+hyperfine -w 2 '../../target/release/arachne < main.arc' 'python3 main.py'