--- a/vm/src/decode.rs
+++ b/vm/src/decode.rs
@@ -1,11 +1,11 @@
 use core::mem::size_of;
 
-#[inline(always)]
+#[inline]
 pub fn decode_f64(codes: &[u8], index: &mut usize) -> f64 {
     f64::from_bits(decode_u64(codes, index))
 }
 
-#[inline(always)]
+#[inline]
 pub fn decode_u64(codes: &[u8], index: &mut usize) -> u64 {
     const SIZE: usize = size_of::<u64>();
     let mut bytes = [0u8; SIZE];
@@ -17,7 +17,7 @@ pub fn decode_u64(codes: &[u8], index: &mut usize) -> u64 {
     u64::from_le_bytes(bytes)
 }
 
-#[inline(always)]
+#[inline]
 pub fn decode_u32(codes: &[u8], index: &mut usize) -> u32 {
     const SIZE: usize = size_of::<u32>();
     let mut bytes = [0u8; SIZE];
@@ -29,7 +29,7 @@ pub fn decode_u32(codes: &[u8], index: &mut usize) -> u32 {
     u32::from_le_bytes(bytes)
 }
 
-#[inline(always)]
+#[inline]
 pub fn decode_u16(codes: &[u8], index: &mut usize) -> u16 {
     const SIZE: usize = size_of::<u16>();
     let mut bytes = [0u8; SIZE];
@@ -41,7 +41,7 @@ pub fn decode_u16(codes: &[u8], index: &mut usize) -> u16 {
     u16::from_le_bytes(bytes)
 }
 
-#[inline(always)]
+#[inline]
 pub fn decode_u8(codes: &[u8], index: &mut usize) -> u8 {
     let value = codes[*index];
 
@@ -50,7 +50,7 @@ pub fn decode_u8(codes: &[u8], index: &mut usize) -> u8 {
     value
 }
 
-#[inline(always)]
+#[inline]
 pub fn decode_bytes<'a>(codes: &'a [u8], len: usize, index: &mut usize) -> &'a [u8] {
     let value = &codes[*index..*index + len];