# Benchmarks

Before:

```sh
> tools/bench.sh
    Finished release [optimized] target(s) in 0.06s
    Updating crates.io index
     Ignored package `hyperfine v1.17.0` is already installed, use --force to override
Benchmark 1: ../../target/release/arachne < main.arc
  Time (mean ± σ):     18.394 s ±  0.210 s    [User: 18.261 s, System: 0.059 s]
  Range (min … max):   18.139 s … 18.750 s    10 runs

Benchmark 2: python3 main.py
  Time (mean ± σ):     12.092 s ±  0.114 s    [User: 12.034 s, System: 0.045 s]
  Range (min … max):   11.953 s … 12.282 s    10 runs

Summary
  python3 main.py ran
    1.52 ± 0.02 times faster than ../../target/release/arachne < main.arc
Benchmark 1: ../../target/release/arachne < main.arc
  Time (mean ± σ):      4.438 s ±  0.063 s    [User: 4.402 s, System: 0.015 s]
  Range (min … max):    4.347 s …  4.548 s    10 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 2: python3 main.py
  Time (mean ± σ):      2.347 s ±  0.049 s    [User: 2.325 s, System: 0.013 s]
  Range (min … max):    2.278 s …  2.451 s    10 runs

Summary
  python3 main.py ran
    1.89 ± 0.05 times faster than ../../target/release/arachne < main.arc
```

After:

```sh
> tools/bench.sh
    Finished release [optimized] target(s) in 0.04s
    Updating crates.io index
     Ignored package `hyperfine v1.17.0` is already installed, use --force to override
Benchmark 1: ../../target/release/arachne < main.arc
  Time (mean ± σ):     12.134 s ±  0.122 s    [User: 12.045 s, System: 0.052 s]
  Range (min … max):   12.029 s … 12.417 s    10 runs

Benchmark 2: python3 main.py
  Time (mean ± σ):     11.780 s ±  0.080 s    [User: 11.711 s, System: 0.056 s]
  Range (min … max):   11.647 s … 11.881 s    10 runs

Summary
  python3 main.py ran
    1.03 ± 0.01 times faster than ../../target/release/arachne < main.arc
Benchmark 1: ../../target/release/arachne < main.arc
  Time (mean ± σ):      3.260 s ±  0.030 s    [User: 3.244 s, System: 0.015 s]
  Range (min … max):    3.220 s …  3.302 s    10 runs

Benchmark 2: python3 main.py
  Time (mean ± σ):      2.296 s ±  0.020 s    [User: 2.278 s, System: 0.016 s]
  Range (min … max):    2.274 s …  2.337 s    10 runs

Summary
  python3 main.py ran
    1.42 ± 0.02 times faster than ../../target/release/arachne < main.arc
```